+++ extension.js
@@ -5,11 +5,11 @@
 // Import the module and reference it with the alias vscode in your code below
 var vscode = require('vscode');
 var window = vscode.window;
-const sbHL7Version = window.createStatusBarItem(vscode.StatusBarAlignment.Left);
 var workspace = vscode.workspace;
 var hl7Schema;
 var hl7Fields;
-
+// the status bar item to display current HL7 schema this is loaded
+var statusbarHL7Version = window.createStatusBarItem(vscode.StatusBarAlignment.Left);
 // the list of fields to highlight
 var fieldSelectionList = [];
 // the list of fields with hover decorations (displaying the field description);
@@ -19,6 +19,30 @@ var currentDecoration;
 // stores the current hover decorations
 var currentHoverDecoration;
 
+
+// Determine if the file is a HL7 file (returns true/false). 
+// This expects that the file extension is .hl7, or the first line contains
+// a MSH segment (or FHS of BHS segment for batch files).
+function IsHL7File(editor) {
+    if (editor) {
+        if (editor.document.languageId == "hl7") {
+            console.log("HL7 file extension detected");
+            return true;
+        }
+        firstLine = editor.document.lineAt(0).text;
+        var hl7HeaderRegex = /(^MSH\|)|(^FHS\|)|(^BHS\|)/i // new RegExp("(^MSH\|)|(^FHS\|)|(^BHS\|)", 'i');
+        if (hl7HeaderRegex.test(firstLine)) {
+            console.log("HL7 header line detected");
+            return true;
+        }
+        else {
+            return false;
+        }
+    }
+    else {
+        return false;
+    }
+}
 
 // add leading spaces to right pad a string
 function padRight(stringToPad, padLength) {
@@ -33,13 +57,13 @@ function padRight(stringToPad, padLength) {
 }
 
 // mask out the nominated component from the field string. 
-// if no component is nomintated, mask all components.
+// if no component is nominated, mask all components.
 // Assumes a field string includes components delimited by '^'
 function maskComponent(fieldToMask, componentNumber) {
     var returnField = "";
     var components = fieldToMask.split('^');
 
-    // no component specified, masks all components and join back into a field string from the modified compoenents.
+    // no component specified, masks all components and join back into a field string from the modified components.
     if (!componentNumber) {
         for (componentIndex = 0; componentIndex < components.length; componentIndex++) {
             components[componentIndex] = components[componentIndex].replace(/\w/g, '#')
@@ -61,7 +85,7 @@ function maskComponent(fieldToMask, componentNumber) {
 }
 
 // Mask all items in a single field, including repeating items.
-// optionally limit the mask to a specific compoent of the field
+// optionally limit the mask to a specific component of the field
 function maskField(fieldToMask, componentNumber) {
     // mask out mother's maiden name
     var fieldRepeats = fieldToMask.split('~')
@@ -72,7 +96,7 @@ function maskField(fieldToMask, componentNumber) {
     return fieldRepeats;
 }
 
-// Mask all fields in an array of fields. Optionaly start masking fields occuring from startingFieldPossition (1 based index of fields)  
+// Mask all fields in an array of fields. Optionally start masking fields occurring from startingFieldPosition (1 based index of fields)  
 function maskFieldList(fieldListToMask, startingPosition) {
     if (!startingPosition) {
         startingPosition = 1;
@@ -102,7 +126,7 @@ function GetFieldIndex(hl7ItemlocationString) {
 // return the unique names of all segments in the message. Return as a associative array indexed by segment name. key values are not consequential.
 function GetAllSegmentNames(document) {
     var segmentHashtable = {};
-    var segmentRegex = new RegExp("^[A-Z]{2}([A-Z]|[0-9])\|", 'i');
+    var segmentRegex = /^[A-Z]{2}([A-Z]|[0-9])\|/i
     for (var i = 0; i < document.lineCount; i++) {
         var currentSegment = document.lineAt(i).text;
         if (segmentRegex.test(currentSegment)) {
@@ -140,34 +164,36 @@ function LoadHL7Schema() {
             else {
                 console.log("Schema for HL7 version " + hl7Version + " is not supported. Defaulting to v2.7.1 schema");
                 hl7Version = "2.7.1";
-                hl7SchemaTooltip = "HL7 version not detected. Defaulting to v" + hl7Version; 
+                hl7SchemaTooltip = "HL7 version not detected. Defaulting to v" + hl7Version;
                 hl7Schema = require('./schema/2.7.1/segments.js');
                 hl7Fields = require('./schema/2.7.1/fields.js');
             }
             // show HL7 version in status bar
-            sbHL7Version.color = 'white';
-            sbHL7Version.text = "$(info) HL7 schema: v" + hl7Version;  // $(info) - GitHub Octicon - https://octicons.github.com/
-            sbHL7Version.tooltip = hl7SchemaTooltip;
-            sbHL7Version.show();
+            statusbarHL7Version.color = 'white';
+            statusbarHL7Version.text = "$(info) HL7 schema: v" + hl7Version;  // $(info) - GitHub Octicon - https://octicons.github.com/
+            statusbarHL7Version.tooltip = hl7SchemaTooltip;
+            statusbarHL7Version.show();
         }
         // if the first line is not a MSH segment (this would be unexpected), default to the 2.7.1 schema
         else {
             hl7Schema = require('./schema/2.7.1/segments.js');
             hl7Fields = require('./schema/2.7.1/fields.js');
             console.log("HL7 version could not be detected. Defaulting to v2.7.1 schema.");
+            statusbarHL7Version.hide();
         }
     }
 }
 
 // this method is called when the extension is activated
 function activate(context) {
-
-    // Use the console to output diagnostic information (console.log) and errors (console.error)
-    // This line of code will only be executed once when your extension is activated
     console.log('The extension "hl7tools" is now active.');
-
-    // exit if the editor is not active
     var activeEditor = window.activeTextEditor
+    // only activate the field descriptions if it is identified as a HL7 file  
+    if (!IsHL7File(activeEditor)) {
+        statusbarHL7Version.hide();
+        return;
+    }
+    // exit if the editor is not active
     if (!activeEditor) {
         return;
     }
@@ -180,27 +206,38 @@ function activate(context) {
 
     // the active document has changed. 
     window.onDidChangeActiveTextEditor(function (editor) {
-        activeEditor = editor;
         if (editor) {
-            // the new document may be a different version of HL7, so load the approprate version of schema
-            LoadHL7Schema();
-            UpdateFieldDescriptions();
+            // only activate the field descriptions if it is identified as a HL7 file  
+            if (IsHL7File(editor)) {
+                // the new document may be a different version of HL7, so load the appropriate version of schema
+                LoadHL7Schema();
+                UpdateFieldDescriptions();
+            }
+            else {
+                statusbarHL7Version.hide();
+            }
         }
     }, null, context.subscriptions);
 
     // document text has changed
     workspace.onDidChangeTextDocument(function (event) {
-        if (activeEditor && event.document === activeEditor.document) {
-            UpdateFieldDescriptions();
+        if (activeEditor && (event.document === activeEditor.document)) {
+            // only activate the field descriptions if it is identified as a HL7 file  
+            if (IsHL7File(editor)) {
+                UpdateFieldDescriptions();
+            }
+            else {
+                statusbarHL7Version.hide();
+            }
         }
     }, null, context.subscriptions);
 
     //-------------------------------------------------------------------------------------------
-    // this function highlights HL7 items in the message based on item possition identified by user.
+    // this function highlights HL7 items in the message based on item position identified by user.
     var highlightFieldCommand = vscode.commands.registerCommand('hl7tools.HighlightHL7Item', function () {
         console.log('In function Highlight Field');
 
-        // associative array indexed by segmentname, with the field index as a value. 
+        // associative array indexed by segment name, with the field index as a value. 
         var locationHashtable = {};
 
         // exit if the editor is not active
@@ -243,7 +280,7 @@ function activate(context) {
             }
             // else assume the user has provided a field description to search for.
             else {
-                // find mathcing field names for any segment present in the message
+                // find matching field names for any segment present in the message
                 var segmentHash = GetAllSegmentNames(currentDoc);
                 for (var key in segmentHash) {
                     var segmentDef = hl7Schema[key];
@@ -277,7 +314,7 @@ function activate(context) {
                     var fieldCount = 1;
                     // get the location of field delimiter characters
                     while (match = regEx.exec(currentLine)) {
-                        // if the start position was located in the previous iteration, then this must be the end posssition
+                        // if the start position was located in the previous iteration, then this must be the end position
                         if (startPos) {
                             endPos = activeEditor.document.positionAt(positionOffset + match.index);
                             var decoration = { range: new vscode.Range(startPos, endPos) };
@@ -342,7 +379,7 @@ function activate(context) {
                     patientIDList[i] = maskComponent(patientIDList[i]);
                 }
                 fields[3] = patientIDList.join('~');
-                // mask out specific PID fields contined in the array below (1 based index - e.g. 4 = PID-4). fields[0] is the segment name.
+                // mask out specific PID fields continued in the array below (1 based index - e.g. 4 = PID-4). fields[0] is the segment name.
                 var pidFieldsToMask = [4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 19, 20, 21, 22, 23, 26, 27, 28];
                 for (i = 0; i < pidFieldsToMask.length; i++) {
                     if (pidFieldsToMask[i] < fields.length) {
@@ -355,7 +392,7 @@ function activate(context) {
             }
             // mask out specific next of kin fields
             else if ((fields[0]).toUpperCase() === "NK1") {
-                // mask out specific PID fields contined in the array below (1 based index - e.g. 4 = PID-4). fields[0] is the segment name.
+                // mask out specific PID fields continued in the array below (1 based index - e.g. 4 = PID-4). fields[0] is the segment name.
                 var nk1FieldsToMask = [2, 4, 5, 6, 7, 10, 11, 12, 13, 14, 15, 16, 19, 20, 25, 26, 27, 28, 29, 30, 31, 32, 33, 35, 37, 38];
                 for (i = 0; i < nk1FieldsToMask.length; i++) {
                     if (nk1FieldsToMask[i] < fields.length) {
@@ -422,6 +459,7 @@ function activate(context) {
     // The function is based on TokenizeLine from https://github.com/pagebrooks/vscode-hl7 . Modified to 
     // support repeating fields and make field indexes start at 1 (instead of 0) to match the HL7 field naming scheme. 
     var displaySegmentCommand = vscode.commands.registerCommand('hl7tools.DisplaySegmentFields', function () {
+
         console.log('In function DisplaySegmentFields');
 
         // exit if the editor is not active
@@ -509,7 +547,7 @@ function activate(context) {
             }
             else {
                 for (var j = 0; j < output[i].values.length; j++) {
-                    // create unicode 'border' charcters for components of fields
+                    // create unicode 'border' characters for components of fields
                     var border = "├";
                     if (j == (output[i].values.length - 1)) {
                         border = "└";
@@ -528,7 +566,7 @@ function activate(context) {
                         value += output[i].values[j];
                     }
 
-                    // include the repeat number for repeating fields. e.g. PID-3[2].1 would be the first componennt of the second repeat of the PID-3 field. 
+                    // include the repeat number for repeating fields. e.g. PID-3[2].1 would be the first component of the second repeat of the PID-3 field. 
                     else {
                         value += padRight('\n ' + border + ' ' + output[i].segment + '[' + output[i].repeat.toString() + '].' + (j + 1) + ' (' + componentDescription + ') ', prefix.length + 1);
                         value += output[i].values[j];
@@ -546,6 +584,58 @@ function activate(context) {
 
     });
     context.subscriptions.push(displaySegmentCommand);
+
+
+    //-------------------------------------------------------------------------------------------
+    // this function splits HL7 batch files into a separate file per message
+    var splitBatchFileCommand = vscode.commands.registerCommand('hl7tools.SplitBatchFile', function () {
+        var activeEditor = vscode.window.activeTextEditor;
+        if (!activeEditor) {
+            return;
+        }
+        // get the end of line char from the config file to append to each line.
+        var config = vscode.workspace.getConfiguration();
+        var endOfLineChar = config.files.eol;
+
+        var newMessage = "";
+        var batchHeaderRegEx = /(^FHS\|)|(^BHS\|)|(^BTS\|)|(^FTS\|)/i;
+        var mshRegEx = /^MSH\|/i;
+        var currentDoc = activeEditor.document;
+        for (lineIndex = 0; lineIndex < currentDoc.lineCount; lineIndex++) {
+            var currentLine = currentDoc.lineAt(lineIndex).text;
+            // ignore batch header segments (FHS, BHS, BTS, FTS)
+            if (batchHeaderRegEx.test(currentLine)) {
+                continue;
+            }
+            // split the message on MSH segments
+            if (mshRegEx.test(currentLine)) {
+                if (newMessage.length > 0) {
+                    // open the message in a new document, user will be prompted to save on exit
+                    vscode.workspace.openTextDocument({ content: newMessage, language: "hl7" }).then((newDocument) => {
+                        vscode.window.showTextDocument(newDocument, 1, false).then(e => {
+                        });
+                    }, (error) => {
+                        console.error(error);
+                    });
+                }
+                newMessage = currentLine;
+            }
+            else {
+                newMessage += endOfLineChar + currentLine
+            }
+        }
+        // write the last message to a new document 
+        if (newMessage.length > 0) {
+            vscode.workspace.openTextDocument({ content: newMessage, language: "hl7" }).then((newDocument) => {
+                vscode.window.showTextDocument(newDocument, 1, false).then(e => {
+                });
+            }, (error) => {
+                console.error(error);
+            });
+        }
+    });
+
+    context.subscriptions.push(splitBatchFileCommand);
 
     //-------------------------------------------------------------------------------------------
     // apply descriptions to each field as a hover decoration (tooltip)
@@ -584,12 +674,12 @@ function activate(context) {
             var fieldCount = -1;
             var previousEndPos = null;
             var fieldDescription = "";
-            // ignore all lines that do not at least contain a segmentname and field delimeter. This should be the absolut minimum for a segment
+            // ignore all lines that do not at least contain a segment name and field delimiter. This should be the absolute minimum for a segment
             if (!validSegmentRegEx.test(currentLine)) {
                 positionOffset += currentLine.length + endOfLineLength;
                 continue;
             }
-            // the first delimeter is a field for MSH segments
+            // the first delimiter is a field for MSH segments
             if (segmentName.toUpperCase() == "MSH") {
                 fieldCount++;
             }
@@ -612,7 +702,7 @@ function activate(context) {
                 }
                 fieldCount++;
             }
-            // add a decoration for the last field in the segment (not bounded by a field delimeter) 
+            // add a decoration for the last field in the segment (not bounded by a field delimiter) 
             startPos = previousEndPos;
             endPos = activeEditor.document.positionAt(positionOffset + (currentLine.length + 1));
             try {